requirements.txt
# absl-py==1.4.0
# aiosignal==1.3.1
# appdirs==1.4.4
# attrs==23.1.0
# cachetools==5.3.1
# certifi==2023.5.7
# cffi==1.15.1
# charset-normalizer==3.1.0
# click==8.1.3
# cloudpickle==2.2.1
# cmake==3.26.4
# contourpy==1.1.0
# cycler==0.11.0
# Cython==0.29.35
# d3rlpy==0.1
# d4rl==1.1
# dm-control==1.0.12
# dm-env==1.6
# dm-tree==0.1.8
# docker-pycreds==0.4.0
# Farama-Notifications==0.0.4
# fasteners==0.18
# filelock==3.12.2
# fonttools==4.40.0
# frozenlist==1.3.3
# fsspec==2023.4.0
# funcsigs==1.0.2
# gitdb==4.0.11
# GitPython==3.1.41
# glfw==2.5.9
# google-auth==2.20.0
# google-auth-oauthlib==1.0.0
# GPUtil==1.4.0
# grpcio==1.51.3
# gtimer==1.0.0b5
# gym==0.23.1
# gym-notices==0.0.8
# gymnasium==0.29.1
# h5py==3.9.0
# idna==3.4
# imageio==2.31.1
# importlib-metadata==6.6.0
# importlib-resources==5.12.0
# Jinja2==3.1.2
# joblib==1.3.2
# jsonschema==4.17.3
# kiwisolver==1.4.4
# labmaze==1.0.6
# lazy_loader==0.3
# lit==16.0.6
# lxml==4.9.2
# Markdown==3.4.3
# MarkupSafe==2.1.3
# matplotlib==3.7.1
# # mjrl==1.0.0
# mpi4py==3.0.3
# mpmath==1.3.0
# msgpack==1.0.5
# # mujoco==2.3.5
# # mujoco-py==2.1.2.14
# networkx==3.1
# numpy==1.24.3
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
# oauthlib==3.2.2
# opencv-python==4.9.0.80
# packaging==23.1
# pandas==2.0.2
# Pillow==9.5.0
# pkgutil_resolve_name==1.3.10
# protobuf==4.23.3
# psutil==5.9.8
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pybullet==3.2.5
# pycparser==2.21
# PyOpenGL==3.1.7
# pyparsing==3.0.9
# pyrsistent==0.19.3
# python-dateutil==2.8.2
# pytz==2023.3
# PyWavelets==1.4.1
# PyYAML==6.0
# ray==2.5.0
# requests==2.31.0
# requests-oauthlib==1.3.1
# rsa==4.9
# scikit-image==0.21.0
# scikit-learn==1.3.2
# scipy==1.10.1
# sentry-sdk==1.39.2
# setproctitle==1.3.3
# six==1.16.0
# smmap==5.0.1
# stable-baselines3==2.2.1
# structlog==23.3.0
# sympy==1.12
# tensorboard==2.13.0
# tensorboard-data-server==0.7.1
# tensorboardX==2.6.2.2
# termcolor==2.3.0
# threadpoolctl==3.2.0
# tifffile==2023.7.10
# # torch==2.1.1+cu118
# # torch-tb-profiler==0.4.3
# # torch-truncnorm==0.0.2
# # torchaudio==2.1.1+cu118
# # torchvision==0.16.2+cu118
# tqdm==4.65.0
# triton==2.1.0
# typing_extensions==4.6.3
# tzdata==2023.3
# urllib3==1.26.16
# wandb==0.16.2
# Werkzeug==2.3.6
# zipp==3.15.0
absl-py==1.4.0
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
d3rlpy==0.1
d4rl==1.1
dm-control==1.0.12
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
fasteners==0.18
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.4.0
funcsigs==1.0.2
gitdb==4.0.11
GitPython==3.1.41
glfw==2.5.9
google-auth==2.20.0
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
grpcio==1.51.3
gtimer==1.0.0b5
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.9.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.4
labmaze==1.0.6
lazy_loader==0.3
lit==16.0.6
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
# mjrl==1.0.0
mpi4py==3.0.3
mpmath==1.3.0
msgpack==1.0.5
# mujoco==2.3.5
# mujoco-py==2.1.2.14
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.23.3
psutil==5.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycparser==2.21
PyOpenGL==3.1.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
ray==2.5.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stable-baselines3==2.2.1
structlog==23.3.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
